:adhesive_bandage: 변수명, 파일 이름, for 문 안의 코드 등 간단한 수정
